# server/docs/swagger-bundled.yaml
openapi: 3.0.0
info:
  title: Users API
  version: 1.0.0
  description: API for managing users with HAL-style hypermedia responses
servers:
  - url: 'http://localhost:3001'
paths:
...
  /api/auctions:
    get:
      summary: Get all auctions
      description: 'Retrieve a list of all auctions with optional pagination, sorting, and filtering.'
      tags:
        - All Auctions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Number of auctions to retrieve
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of auctions to skip
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: end_time
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - ASC
              - DESC
            default: ASC
          description: Sort order (ascending or descending)
      responses:
        '200':
          description: A list of auctions
          content:
            application/json:
              schema:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/auctions/
                      profile:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /profile/auctions/
                      create:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/auctions
                          method:
                            type: string
                            example: POST
                  _embedded:
                    type: object
                    properties:
                      auctions:
                        type: array
                        items:
                          type: object
                          properties:
                            _links:
                              type: object
                              properties:
                                self:
                                  type: object
                                  properties:
                                    href:
                                      type: string
                                      example: /api/auctions/3
                                delete:
                                  type: object
                                  properties:
                                    href:
                                      type: string
                                      example: /api/auctions/3
                                    method:
                                      type: string
                                      example: DELETE
                                bidsByAuction:
                                  type: object
                                  properties:
                                    href:
                                      type: string
                                      example: /api/auctions/3/bids
                            id:
                              type: integer
                              example: 3
                            description:
                              type: string
                              example: New orchid Pizza with ergonomic design for bogus comfort
                            end_time:
                              type: string
                              format: date-time
                              example: '2025-06-27T09:21:07.282Z'
                            starting_price:
                              type: string
                              example: '740.89'
                            current_price:
                              type: string
                              example: '495.90'
                            item_id:
                              type: integer
                              nullable: true
                              example: 18
                            seller_id:
                              type: integer
                              example: 1
        '204':
          description: No auctions found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: No auctions data found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal Server Error
    post:
      summary: Add a new auction
      description: Create a new auction in the database.
      tags:
        - All Auctions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAuctionRequest'
      responses:
        '201':
          description: Auction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auction'
              example:
                _links:
                  self:
                    href: /api/auctions/9
                  all:
                    href: /api/auctions
                  profile:
                    href: /profile/auctions
                  delete:
                    href: /api/auctions/9
                    method: DELETE
                start_time: '2025-04-16T23:25:17.049Z'
                emaidSend: false
                id: 9
                item_id: 3
                seller_id: 7
                description: A rare collectible item.
                starting_price: '50.00'
                current_price: '50.00'
                end_time: '2023-12-31T23:59:59.000Z'
                itemId: 3
        '400':
          $ref: '#/components/schemas/BadRequest'
        '500':
          $ref: '#/components/schemas/InternalServerError'
...
  /api/categories:
    get:
      summary: Get all categories
      description: Retrieve a list of all categories.
      tags:
        - Categories
      responses:
        '200':
          description: A list of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/categories/
                      profile:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /profile/categories/
                      create:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/categories
                          method:
                            type: string
                            example: POST
                  _embedded:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          type: object
                          allOf:
                            - type: object
                              properties:
                                _links:
                                  type: object
                                  properties:
                                    self:
                                      type: object
                                      properties:
                                        href:
                                          type: string
                                          example: /api/categories/1
                                    delete:
                                      type: object
                                      properties:
                                        href:
                                          type: string
                                          example: /api/categories/1
                                        method:
                                          type: string
                                          example: DELETE
                            - $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/schemas/InternalServerError'
    post:
      summary: Add a new category
      description: Create a new category in the database.
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCategoryRequest'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/categories/1
                      all:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/categories
                      profile:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /profile/categories/
                      delete:
                        type: object
                        properties:
                          href:
                            type: string
                            example: /api/categories/1
                          method:
                            type: string
                            example: DELETE
                  dataValues:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: Electronics
                      description:
                        type: string
                        example: Items related to electronic devices.
                  _previousDataValues:
                    type: object
                    allOf:
                      - $ref: '#/components/schemas/Category'
                  uniqno:
                    type: integer
                    example: 1
                  _changed:
                    type: object
                    example: {}
                  _options:
                    type: object
                    properties:
                      isNewRecord:
                        type: boolean
                        example: true
                      _schema:
                        type: string
                        example: null
                      _schemaDelimiter:
                        type: string
                        example: ''
                  isNewRecord:
                    type: boolean
                    example: false
        '400':
          $ref: '#/components/schemas/BadRequest'
        '409':
          description: Conflict - Category already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Category already exists
        '500':
          $ref: '#/components/schemas/InternalServerError'
...
  /profile/categories:
    get:
      tags:
        - Profile
      summary: Retrieve category profile information
      description: Fetches the profile details and actions for categories.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesProfile'
        '404':
          description: Categories Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Categories Profile not found
...
  /profile/auctions:
    get:
      tags:
        - Profile
      summary: Retrieve auction profile information
      description: Fetches the profile details and actions for auctions.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionsProfile'
        '404':
          description: Auctions Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Auctions Profile not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Auction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Vintage Watch
        description:
          type: string
          example: A rare vintage watch in excellent condition.
        starting_price:
          type: number
          format: float
          example: 100
        current_price:
          type: number
          format: float
          example: 100
        item_id:
          type: integer
          example: 101
        seller_id:
          type: integer
          example: 1
        end_time:
          type: string
          format: date-time
          example: '2023-12-31T23:59:59Z'
        created_at:
          type: string
          format: date-time
          example: '2023-01-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-01-02T12:00:00Z'
    AddAuctionRequest:
      type: object
      required:
        - item_id
        - description
        - starting_price
        - end_time
      properties:
        item_id:
          type: integer
          example: 3
        description:
          type: string
          example: A rare collectible item.
        starting_price:
          type: number
          format: float
          example: 50
        end_time:
          type: string
          format: date-time
          example: '2023-12-31T23:59:59Z'
...
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        description:
          type: string
          example: Items related to electronic devices.
        name:
          type: string
          example: Electronics
    AddCategoryRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: Electronics
        description:
          type: string
          example: Items related to electronic devices.
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request body
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
...
    CategoriesProfile:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
                  example: /profile/categories
            item:
              type: object
              properties:
                href:
                  type: string
                  example: '/api/categories/:id'
                templated:
                  type: boolean
                  example: true
        properties:
          type: object
          properties:
            id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The unique identifier for the category.
            description:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: A description of the category.
            name:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: The name of the category.
                maxLength:
                  type: integer
                  example: 100
        actions:
          type: array
          items:
            oneOf:
              - type: object
                properties:
                  name:
                    type: string
                    example: create
                  method:
                    type: string
                    example: POST
                  href:
                    type: string
                    example: /api/categories
                  title:
                    type: string
                    example: Create a new object to categories
              - type: object
                properties:
                  name:
                    type: string
                    example: get all
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: /api/categories
                  title:
                    type: string
                    example: Get all objects from categories
              - type: object
                properties:
                  name:
                    type: string
                    example: get single
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: '/api/categories/:id'
                  title:
                    type: string
                    example: 'Get a single object from categories by :id'
              - type: object
                properties:
                  name:
                    type: string
                    example: delete
                  method:
                    type: string
                    example: DELETE
                  href:
                    type: string
                    example: '/api/categories/:id'
                  templated:
                    type: boolean
                    example: true
                  title:
                    type: string
                    example: Delete object from categories
    AuctionsProfile:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
                  example: /profile/auctions
            item:
              type: object
              properties:
                href:
                  type: string
                  example: '/api/auctions/:id'
                templated:
                  type: boolean
                  example: true
        properties:
          type: object
          properties:
            id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The unique identifier for the auction.
            description:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: A detailed description of the auction item.
                maxLength:
                  type: integer
                  example: 1000
            item_id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The ID of the item being auctioned.
                format:
                  type: string
                  example: int64
            seller_id:
              type: object
              properties:
                type:
                  type: string
                  example: integer
                description:
                  type: string
                  example: The ID of the seller who created the auction.
                format:
                  type: string
                  example: int64
            end_time:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: The date and time when the auction ends.
                format:
                  type: string
                  example: date-time
            starting_price:
              type: object
              properties:
                type:
                  type: string
                  example: number
                description:
                  type: string
                  example: The starting price of the auction item.
                format:
                  type: string
                  example: decimal
                minimum:
                  type: number
                  example: 0
            current_price:
              type: object
              properties:
                type:
                  type: string
                  example: number
                description:
                  type: string
                  example: The current price of the auction item.
                format:
                  type: string
                  example: decimal
                minimum:
                  type: number
                  example: 0
            created_at:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: The date and time when the auction was created.
                format:
                  type: string
                  example: date-time
            updated_at:
              type: object
              properties:
                type:
                  type: string
                  example: string
                description:
                  type: string
                  example: The date and time when the auction was last updated.
                format:
                  type: string
                  example: date-time
        actions:
          type: array
          items:
            oneOf:
              - type: object
                properties:
                  name:
                    type: string
                    example: create
                  method:
                    type: string
                    example: POST
                  href:
                    type: string
                    example: /api/auctions
                  title:
                    type: string
                    example: Create a new object to auctions
              - type: object
                properties:
                  name:
                    type: string
                    example: get all
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: /api/auctions
                  title:
                    type: string
                    example: Get all objects from auctions
              - type: object
                properties:
                  name:
                    type: string
                    example: get single
                  method:
                    type: string
                    example: GET
                  href:
                    type: string
                    example: '/api/auctions/:id'
                  title:
                    type: string
                    example: 'Get a single object from auctions by :id'
              - type: object
                properties:
                  name:
                    type: string
                    example: delete
                  method:
                    type: string
                    example: DELETE
                  href:
                    type: string
                    example: '/api/auctions/:id'
                  templated:
                    type: boolean
                    example: true
                  title:
                    type: string
                    example: Delete object from auctions
...